renderer refactor
