+++ crates/vulkan-renderer/Cargo.toml
@@ -5,8 +5,10 @@ edition = "2021"
 
 [dependencies]
 vulkan-instance = { workspace = true }
+hdr-capture = { workspace = true }
 
 tracing = { workspace = true }
 thiserror = { workspace = true }
 winit = { workspace = true }
 ash = { workspace = true }
+bytemuck = { workspace = true }
