+++ crates/vulkan-renderer/src/objects/mod.rs
@@ -1,0 +1,9 @@
+mod border;
+mod capture;
+mod mouse_guides;
+mod selection;
+
+pub use border::Border;
+pub use capture::Capture;
+pub use mouse_guides::MouseGuides;
+pub use selection::Selection;
