+++ crates/vulkan-renderer/src/pipelines/border.rs
@@ -1,19 +1,15 @@
-use std::{io::Cursor, mem::offset_of};
+use std::mem::{offset_of, size_of};
 
-use ash::{
-    util::read_spv,
-    vk::{
-        BlendFactor, BlendOp, ColorComponentFlags, Format, Pipeline,
-        PipelineColorBlendAttachmentState, PipelineLayout, PipelineLayoutCreateInfo,
-        PipelineRenderingCreateInfo, PipelineShaderStageCreateInfo,
-        PipelineVertexInputStateCreateInfo, PushConstantRange, ShaderModule,
-        ShaderModuleCreateInfo, ShaderStageFlags, VertexInputAttributeDescription,
-        VertexInputBindingDescription, VertexInputRate, Viewport,
-    },
+use ash::vk::{
+    BlendFactor, BlendOp, ColorComponentFlags, Format, Pipeline, PipelineColorBlendAttachmentState,
+    PipelineLayout, PipelineLayoutCreateInfo, PipelineRenderingCreateInfo,
+    PipelineShaderStageCreateInfo, PipelineVertexInputStateCreateInfo, PushConstantRange,
+    ShaderModule, ShaderStageFlags, VertexInputAttributeDescription, VertexInputBindingDescription,
+    VertexInputRate, Viewport,
 };
-use vulkan_instance::VulkanInstance;
-
-use super::Error;
+use bytemuck::{Pod, Zeroable};
+use tracing::instrument;
+use vulkan_instance::{VulkanError, VulkanInstance};
 
 #[derive(Clone, Copy, Debug)]
 pub struct Vertex {
@@ -22,11 +18,22 @@ pub struct Vertex {
     pub flags: u32,
 }
 
+#[derive(Clone, Copy, Debug, Pod, Zeroable)]
+#[repr(C)]
+pub struct PushConstants {
+    pub base_position: [f32; 2],
+    pub base_size: [f32; 2],
+    pub target_position: [f32; 2],
+    pub target_size: [f32; 2],
+    pub line_size: [f32; 2],
+}
+
+#[instrument("boder::create_pipeline", skip_all, err)]
 pub fn create_pipeline(
     vk: &VulkanInstance,
     pipeline_rendering_create_info: PipelineRenderingCreateInfo,
     viewport: Viewport,
-) -> Result<(Pipeline, PipelineLayout, [ShaderModule; 2]), Error> {
+) -> Result<(Pipeline, PipelineLayout, [ShaderModule; 2]), VulkanError> {
     let vertex_input_binding_descriptions = [VertexInputBindingDescription {
         binding: 0,
         stride: std::mem::size_of::<Vertex>() as u32,
@@ -69,24 +76,14 @@ pub fn create_pipeline(
     };
 
     let (vs, vs_entry) = unsafe {
-        let mut shader_file = Cursor::new(&include_bytes!("../shaders/border.vert.spv")[..]);
-        let shader_code = read_spv(&mut shader_file).map_err(Error::ReadShader)?;
-        let shader_info = ShaderModuleCreateInfo::default().code(&shader_code);
-        let shader_module = vk
-            .device
-            .create_shader_module(&shader_info, None)
-            .map_err(|e| Error::Vulkan(e, "creating shader module"))?;
+        let shader_module =
+            vk.create_shader_module(include_bytes!("../shaders/border.vert.spv"))?;
         let shader_entry_name = std::ffi::CStr::from_bytes_with_nul_unchecked(b"main\0");
         (shader_module, shader_entry_name)
     };
     let (fs, fs_entry) = unsafe {
-        let mut shader_file = Cursor::new(&include_bytes!("../shaders/border.frag.spv")[..]);
-        let shader_code = read_spv(&mut shader_file).map_err(Error::ReadShader)?;
-        let shader_info = ShaderModuleCreateInfo::default().code(&shader_code);
-        let shader_module = vk
-            .device
-            .create_shader_module(&shader_info, None)
-            .map_err(|e| Error::Vulkan(e, "creating shader module"))?;
+        let shader_module =
+            vk.create_shader_module(include_bytes!("../shaders/border.frag.spv"))?;
         let shader_entry_name = std::ffi::CStr::from_bytes_with_nul_unchecked(b"main\0");
         (shader_module, shader_entry_name)
     };
@@ -104,7 +101,7 @@ pub fn create_pipeline(
     let push_constant_ranges = [PushConstantRange {
         stage_flags: ShaderStageFlags::VERTEX,
         offset: 0,
-        size: 40,
+        size: size_of::<PushConstants>() as u32,
     }];
 
     let pipeline_layout_create_info =
